fix: change idTrail from double to long in dto

+++ natour-backend/src/main/java/com/natour/natour/model/dto/TrailResponseDto.java
@@ -13,7 +13,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 @NoArgsConstructor
 public class TrailResponseDto {
-    private double idTrail;
+    private long idTrail;
     private UserDetailsDto owner;
     private String name;
     private String description;
